Separate headers from request context
Keep header names, typed policy lists, and extraction utilities in headers.ts while moving Shopify request lifecycle behavior to request-context.ts. This gives each module a focused responsibility and makes new header policies typo-safe without changing request-context behavior.
