+++ packages/hydrogen/src/client/client.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 
 import type { CacheInstance } from "../core";
 import { Cache } from "../core/cache";
-import { createShopifyRequestContext, type ShopifyRequestContext } from "../core/headers";
+import { createShopifyRequestContext, type ShopifyRequestContext } from "../core/request-context";
 import { assert } from "../core/test-utils";
 import { gql } from "../graphql";
 import { createStorefrontClient } from "./client";
