+++ packages/hydrogen/src/core/request-routing/handle-shopify-redirects.test.ts
@@ -2,8 +2,8 @@ import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 
 import { createStorefrontClient } from "../../client/client";
 import type { I18nConfig } from "../../client/types";
-import { createShopifyRequestContext } from "../headers";
 import { configureLogging, resetLoggingForTests } from "../logging";
+import { createShopifyRequestContext } from "../request-context";
 import { createShopifyRouteTemplates } from "../standard-routes/index";
 import { assert, createTestLogger } from "../test-utils";
 import { handleShopifyRedirects } from "./handle-shopify-redirects";
