+++ packages/hydrogen/src/core/request-routing/interceptors/url-redirects.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 
 import { createStorefrontClient } from "../../../client/client";
-import { createShopifyRequestContext } from "../../headers";
+import { createShopifyRequestContext } from "../../request-context";
 import { createShopifyRouteTemplates } from "../../standard-routes/index";
 import { assert } from "../../test-utils";
 import type { RedirectOptions } from "../handle-shopify-redirects";
