+++ packages/hydrogen/src/core/request-routing/route-types.ts
@@ -1,5 +1,5 @@
 import type { StorefrontClient } from "../../client";
-import type { ShopifyRequestContext } from "../headers";
+import type { ShopifyRequestContext } from "../request-context";
 
 type Awaitable<T> = T | Promise<T>;
